requirements.txt
absl-py==2.1.0
ai2thor==5.0.0
allenact==0.5.4
allenact-plugins==0.5.4
aniso8601==9.0.1
annotated-types==0.6.0
antlr4-python3-runtime==4.8
anyio==4.2.0
astunparse==1.6.3
attr==0.3.2
attrs==23.2.0
aws-requests-auth==0.4.3
blinker==1.7.0
botocore==1.34.14
box2d-py==2.3.8
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==1.6.0
colour==0.1.5
compress-json==1.0.10
compress-pickle==2.1.0
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
decorator==4.4.2
Deprecated==1.2.14
distro==1.9.0
docker-pycreds==0.4.0
exceptiongroup==1.2.0
fasttext==0.9.2
filelock==3.13.1
Flask==3.0.0
Flask-RESTful==0.3.10
flatbuffers==1.12
fonttools==4.47.2
fsspec==2023.12.2
ftfy==6.1.3
future==0.18.3
gast==0.4.0
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.27.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.60.1
gym==0.17.3
gym-minigrid==1.0.3
gym-notices==0.0.8
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.26.0
hydra-core==1.1.1
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
llvmlite==0.41.1
loguru==0.7.2
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.8.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
networkx==3.2.1
numba==0.58.1
numpy==1.26.3
numpy-quaternion==2023.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
objathor==0.0.5
objaverse==0.1.7
omegaconf==2.1.2
openai==1.7.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
patsy==0.5.6
pickle5==0.0.11
pillow==10.2.0
platformdirs==4.2.1
prior==1.0.3
procthor==0.0.1.dev2
proglog==0.1.10
progressbar2==4.3.2
protobuf==3.20.0
psutil==5.9.8
pyarrow==16.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyGithub==2.1.1
pyglet==1.5.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyquaternion==0.9.9
python-dateutil==2.8.2
python-sat==1.8.dev13
python-utils==3.8.1
python-xlib==0.33
pytz==2023.3.post1
PyYAML==6.0.1
Quaternion==3.5.2.post4
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scipy==1.11.4
sentry-sdk==2.1.1
setproctitle==1.3.3
shapely==2.0.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
sympy==1.12
tabulate==0.9.0
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorboardX
tensorflow
tensorflow-estimator
tensorflow-io-gcs-filesystem
termcolor==2.4.0
tifffile==2024.1.30
torch==1.11.0
torchvision==0.12.0
tqdm==4.66.1
trimesh==4.3.2
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
wandb==0.17.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
